backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
groq==0.13.0
firebase-admin==6.5.0
google-cloud-firestore==2.19.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.28.1
